move new component out of platform and add RTL tests
